Move configs stuff to separate file
